requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.26.0
jupyter==1.0.0
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.1+cu117
torchinfo==1.8.0
transformers==4.33.2
sentencepiece==0.1.99
protobuf==4.24.3
accelerate==0.23.0
peft==0.5.0
aioredis==2.0.1